move : 메인페이지 csr ssr 렌더링 분리
